removed type_system library
